Updated to the newest template version.

+++ TemplateCodeGenerator.Logic/Generator.cs
@@ -5,11 +5,12 @@ namespace TemplateCodeGenerator.Logic
     using System.Collections.Concurrent;
     using System.Text;
     using TemplateCodeGenerator.Logic.Contracts;
+
     public static partial class Generator
     {
         public static IEnumerable<IGeneratedItem> Generate(string solutionPath)
         {
-            ISolutionProperties solutionProperties = Generation.SolutionProperties.Create(solutionPath);
+            ISolutionProperties solutionProperties = SolutionProperties.Create(solutionPath);
 
             return Generate(solutionProperties);
         }
@@ -51,6 +52,22 @@ namespace TemplateCodeGenerator.Logic
             }
             #endregion WebApiApp
 
+            #region MVVMApp
+            if (configuration.QuerySettingValue<bool>(Common.UnitType.MVVM.ToString(), "All", "All", "Generate", "True"))
+            {
+                var generator = new Generation.MVVMGenerator(solutionProperties);
+
+                tasks.Add(Task.Factory.StartNew(() =>
+                {
+                    var generatedItems = new List<IGeneratedItem>();
+
+                    Console.WriteLine("Create MVVM-Components...");
+                    generatedItems.AddRange(generator.GenerateAll());
+                    result.AddRangeSafe(generatedItems);
+                }));
+            }
+            #endregion MVVMApp
+
             #region AspMvcApp
             if (configuration.QuerySettingValue<bool>(Common.UnitType.AspMvc.ToString(), "All", "All", "Generate", "True"))
             {
@@ -67,21 +84,21 @@ namespace TemplateCodeGenerator.Logic
             }
             #endregion AspMvcApp
 
-            #region MVVMApp
-            if (configuration.QuerySettingValue<bool>(Common.UnitType.MVVM.ToString(), "All", "All", "Generate", "True"))
+            #region ClientBlazorApp
+            if (configuration.QuerySettingValue<bool>(Common.UnitType.ClientBlazor.ToString(), "All", "All", "Generate", "True"))
             {
-                var generator = new Generation.MVVMGenerator(solutionProperties);
+                var generator = new Generation.ClientBlazorGenerator(solutionProperties);
 
                 tasks.Add(Task.Factory.StartNew(() =>
                 {
                     var generatedItems = new List<IGeneratedItem>();
 
-                    Console.WriteLine("Create MVVM-Components...");
+                    Console.WriteLine("Create Client-Blazor-Components...");
                     generatedItems.AddRange(generator.GenerateAll());
                     result.AddRangeSafe(generatedItems);
                 }));
             }
-            #endregion MVVMApp
+            #endregion ClientBlazorApp
 
             #region AngularApp
             if (configuration.QuerySettingValue<bool>(Common.UnitType.Angular.ToString(), "All", "All", "Generate", "True"))
@@ -103,27 +120,44 @@ namespace TemplateCodeGenerator.Logic
             return result;
         }
 
-        public static void DeleteGeneratedFiles(string path)
+        public static void DeleteGeneratedFiles(string sourcePath)
         {
+            var solutionProperties = SolutionProperties.Create(sourcePath);
+            var configuration = new Generation.Configuration(solutionProperties);
+
             Console.WriteLine("Delete all generation files...");
 
             foreach (var searchPattern in StaticLiterals.SourceFileExtensions.Split("|"))
             {
-                var deleteFiles = GetGeneratedFiles(path, searchPattern, new[] { StaticLiterals.GeneratedCodeLabel });
+                var deleteFiles = GetGeneratedFiles(sourcePath, searchPattern, new[] { StaticLiterals.GeneratedCodeLabel });
 
                 foreach (var file in deleteFiles)
                 {
-                    if (Generation.FileHandler.IsTypeScriptFile(file))
+                    var isTemplateProjectFile = solutionProperties.IsTemplateProjectFile(file);
+
+                    if (isTemplateProjectFile)
                     {
-                        Generation.FileHandler.SaveAngularCustomParts(file);
+                        var projectName = solutionProperties.GetProjectNameFromFile(file); ;
+                        var defaultValue = configuration.QuerySettingValue(projectName, "All", "All", "Delete", true.ToString());
+                        var canDelete = configuration.QuerySettingValue<bool>(projectName, "File", Path.GetFileName(file), "Delete", defaultValue);
+
+                        if (canDelete)
+                        {
+                            if (Generation.FileHandler.IsTypeScriptFile(file))
+                            {
+                                Generation.FileHandler.SaveAngularCustomParts(file);
+                            }
+
+                            File.Delete(file);
+                        }
                     }
-                    File.Delete(file);
                 }
             }
-            var defines = Preprocessor.ProjectFile.ReadDefinesInProjectFiles(path);
+            var defines = Preprocessor.ProjectFile.ReadDefinesInProjectFiles(sourcePath);
 
             Preprocessor.ProjectFile.SwitchDefine(defines, Preprocessor.ProjectFile.GeneratedCodePrefix, Preprocessor.ProjectFile.OffPostfix);
-            Preprocessor.ProjectFile.WriteDefinesInProjectFiles(path, defines);
+            Preprocessor.ProjectFile.WriteDefinesInProjectFiles(sourcePath, defines);
+            Preprocessor.RazorFile.SetPreprocessorDefineCommentsInRazorFiles(sourcePath, defines);
         }
         private static IEnumerable<string> GetGeneratedFiles(string path, string searchPattern, string[] labels)
         {
